pyproject.toml
[tool.poetry]
name = "repl-nix-geminiphantom"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
email-validator = "^2.2.0"
google-generativeai = "^0.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
